feat: register  funnel 구조 적용
